Move board state and tests into pyramid.
