+++ neuralknight/models/board.py
