+++ requirements.txt
@@ -1,0 +1,34 @@
+attrs==17.4.0
+beautifulsoup4==4.6.0
+coverage==4.5.1
+hupper==1.1
+Jinja2==2.10
+Mako==1.0.7
+MarkupSafe==1.0
+more-itertools==4.1.0
+PasteDeploy==1.5.2
+plaster==1.0
+plaster-pastedeploy==0.5
+pluggy==0.6.0
+py==1.5.3
+Pygments==2.2.0
+pyramid==1.9.1
+pyramid-debugtoolbar==4.4
+pyramid-jinja2==2.7
+pyramid-mako==1.0.2
+pyramid-retry==0.5
+pyramid-tm==2.2
+pytest==3.5.0
+pytest-cov==2.5.1
+repoze.lru==0.7
+six==1.11.0
+SQLAlchemy==1.2.6
+transaction==2.2.1
+translationstring==1.3
+venusian==1.1.0
+waitress==1.1.0
+WebOb==1.8.1
+WebTest==2.0.29
+zope.deprecation==4.3.0
+zope.interface==4.4.3
+zope.sqlalchemy==1.0
